+++ examples/import/e1_log_context.cc
@@ -1,0 +1,6 @@
+import galay.tracing;
+
+int main() {
+    auto span = galay::tracing::startSpan("checkout");
+    GALAY_LOG_INFO("order accepted {}", 123);
+}